% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hdi.R
\name{plot_hdiData}
\alias{plot_hdiData}
\title{Plot Method for HDI Data}
\usage{
plot_hdiData(x)
}
\arguments{
\item{x}{An 'hdiData object}
}
\description{
Generates 4 plots for value, country, index and year range.
}
