% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hdi.R
\name{print_hdiData}
\alias{print_hdiData}
\title{Print Method for HDI Data}
\usage{
print_hdiData(x)
}
\arguments{
\item{x}{An 'hdiData' object}
}
\description{
Prints the HDI data in a user-friendly format.
}
