% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdi.R
\name{read_hdi}
\alias{read_hdi}
\title{Read HDI Data}
\usage{
read_hdi(country_code)
}
\arguments{
\item{country_code}{The name of the country for which to load the data, used to construct the file name.}
}
\value{
A data frame of class 'hdi_data' with properly formatted columns and data types.
}
\description{
This function reads an HDI data file for a specified country, processes the data,
and returns an object of class 'hdi_data'. It ensures that columns are displayed nicely
and variables have the correct class.
}
\examples{
hdi_data <- read_hdi("irl")
}
