% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hdi.R
\name{summary_hdiData}
\alias{summary_hdiData}
\title{Summary Method for HDI Data}
\usage{
summary_hdiData(x)
}
\arguments{
\item{x}{An 'hdiData' object}
}
\value{
A summary of the HDI data values.
}
\description{
Provides a statistical summary of the HDI data.
}
